+++ examplepatches/RookieEnough/De-Vanced/patches/src/main/kotlin/app/morphe/patches/messenger/inbox/Fingerprints.kt
@@ -38,3 +38,8 @@ internal object LoadInboxAdsFingerprint : Fingerprint(
             "InboxAdsItemSupplierImplementation;"
     },
 )
+
+internal object FriendsInboxTrayFingerprint : Fingerprint(
+    returnType = "Z",
+    strings = listOf("com.facebook.messaging.friendsinboxunit.plugins.inboxunit.FriendsInboxUnitKillSwitch"),
+)